+++ ktorfit-gradle-plugin/src/main/java/de/jensklingenberg/ktorfit/gradle/KtorfitCompilerSubPlugin.kt
@@ -36,7 +36,11 @@ internal class KtorfitCompilerSubPlugin : KotlinCompilerPluginSupportPlugin {
     override fun getCompilerPluginId(): String = COMPILER_PLUGIN_ID
 
     override fun isApplicable(kotlinCompilation: KotlinCompilation<*>): Boolean {
-        return true
+        val kotlinVersion = myproject.ktorfitExtension(
+            GRADLE_TASKNAME
+        ).kotlinVersion.get()
+
+        return kotlinVersion != "-"
     }
 
     @OptIn(ExperimentalKotlinGradlePluginApi::class, ExperimentalBuildToolsApi::class)